+++ m-psrap.go
@@ -11,7 +11,7 @@ import (
 )
 
 func (e *Erasure) PartialStripeMultiRecoverPreliminary(
-	fileName string, slowLatency int, options *Options) (
+	fileName string, slowLatency float64, options *Options) (
 	map[string]string, error) {
 	failNum := 0
 	for i := 0; i < e.DiskNum; i++ {
